Make signal number configurable
For some reason making the signal number configurable significantly
improves performance (by a factor of two), for reasons that elude
me. My previous utilization of SIGRTMIN was incorrect, because
SIGRTMIN is actually a function call and not a constant, but I don't
see how this could have caused a perf regression (rather than just
not working at all).